+++ Data/Retail/ExportedCalendarEvents.lua
@@ -23,7 +23,7 @@ end
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-10-21 20-02-56 ]] --
+-- [[ Exported at 2023-10-25 21-25-09 ]] --
 tasks = {
     {N, 141, 236705}, -- Feast of Winter Veil
     {N, 181, 237000}, -- Noblegarden