+++ Data/Retail/ExportedWidgetEvents.lua
@@ -23,12 +23,13 @@ end
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-10-21 20-02-56 ]] --
+-- [[ Exported at 2023-10-25 21-25-09 ]] --
 tasks = {
     {N, 4729, 4240492, addon.L["Community Feast: Starting Soon"]}, -- Community Feast: Starting Soon
     {N, 4731, 4240492, addon.L["Community Feast: Active"]}, -- Community Feast: Active
     {N, 4987, 236387, addon.L["Time Rift: Active"]}, -- Time Rift: Active
     {N, 4992, 236387, addon.L["Time Rift: Starting Soon"]}, -- Time Rift: Starting Soon
+    {N, 5323, 3939983, addon.L["Superbloom: Active"]}, -- Superbloom: Active
 };
 
 function exportedWidgetEvents.LoadCategories(e)
@@ -36,6 +37,7 @@ function exportedWidgetEvents.LoadCategories(e)
     e[4731].Category = addon.Tabs["Events"].Categories[15].Children[3]; -- Community Feast
     e[4987].Category = addon.Tabs["Events"].Categories[15].Children[6]; -- Time Rifts
     e[4992].Category = addon.Tabs["Events"].Categories[15].Children[6]; -- Time Rifts
+    e[5323].Category = addon.Tabs["Events"].Categories[15].Children[8]; -- Superbloom
 end
 
 function exportedWidgetEvents.InjectDynamicOptions()
@@ -45,4 +47,6 @@ function exportedWidgetEvents.InjectDynamicOptions()
     KrowiAF_RegisterEventOptions("Widget", "Time_Rifts", addon.L["Time Rifts"], {4987}, addon.L["Time Rift: Active"]);
     KrowiAF_RegisterEventOptions("Widget", "Time_Rifts", addon.L["Time Rifts"], {4992}, addon.L["Time Rift: Starting Soon"]);
     KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Time_Rifts", { 4987, 4992 });
+    KrowiAF_RegisterEventOptions("Widget", "Superbloom", addon.L["Superbloom"], {5323}, addon.L["Superbloom: Active"]);
+    KrowiAF_RegisterDeSelectAllEventOptions("Widget", "Superbloom", { 5323 });
 end