+++ Data/WrathClassic/ExportedCategories.lua
@@ -54,7 +54,7 @@ end
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-10-19 07-06-13 ]] --
+-- [[ Exported at 2023-10-25 21-24-55 ]] --
 tasks = {
     {F, 857, "ExcludedCategories"},
     {C, 971, 857},