+++ Localization/enUS.WrathClassic.lua
@@ -1,8 +1,8 @@
-local addonName, addon = ...;
+local _, addon = ...;
 if not addon.Util.IsWrathClassic then
     return;
 end
-local L = LibStub(addon.Libs.AceLocale):NewLocale(addonName, "enUS", true, true);
+local L = addon.Localization.GetDefaultLocale();
 
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --